feat(types)!: separate partial information about channel into own struct
The rationale for this is to make it so that other clients (such ass clients for Leap) can reuse this library without needing to duplicate parts that are used in different aspects of the API. This is a breaking change for people creating their own Channel instances. However, the scope of this should be limited.